--- csv/FEDEX.csv
@@ -1,9 +1,0 @@
-2334734,356.4,FEDEX GUADALAJARA PARK - Calle Teresa Reynoso 4494,,
-2334735,222.75,FEDEX GUADALAJARA PARK - Circuito de los Almendros 30,,
-2334738,285.12,FEDEX GUADALAJARA PARK - Av. Europa del Oeste,,
-2334740,213.84,FEDEX GUADALAJARA PARK - Circuito Artesanos 209,,
-2334741,356.4,FEDEX GUADALAJARA PARK - Austria 840,,
-2334743,213.84,FEDEX GUADALAJARA PARK - Circuito Artesanos 209,,
-2334744,222.75,FEDEX GUADALAJARA PARK - Circuito de los Almendros 30,,
-2334745,231.66,FEDEX GUADALAJARA PARK - Calle Teresa Reynoso 4494,,
-2334746,285.12,FEDEX GUADALAJARA PARK - Av. Europa del Oeste,,
